requirements/dev.txt
-r base.txt

pre-commit==3.5.0
psycopg[binary]==3.1.13
django-stubs[compatible-mypy]==4.2.6
